+++ app/App.js
@@ -5,18 +5,18 @@ import React, { useContext } from 'react';
 import { StyleSheet, useWindowDimensions } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import CustomDrawerContent from '../components/CustomDrawerContent';
-import Almacenes from '../components/views/Almacenes';
+import Almacenes from '../components/views/Almacenes/Almacenes';
 import Entradas from '../components/views/Entradas/Entradas';
 import Productos from '../components/views/Insumos/Productos';
 import Platos from '../components/views/Platos/Platos';
-import Proveedores from '../components/views/Proveedores';
-import Reportes from '../components/views/Reportes';
-import Salidas from '../components/views/Salidas';
-import Usuarios from '../components/views/Usuarios';
+import Proveedores from '../components/views/Proveedores/Proveedores';
+import Reportes from '../components/views/Reportes/Reportes';
+import Salidas from '../components/views/Salidas/Salidas';
+import Usuarios from '../components/views/Usuarios/Usuarios';
 import { AuthContext, AuthProvider } from '../src/AuthContext';
-import HomeScreen from './views/HomeScreen';
-import LoadingApp from './views/LoadingApp';
-import LoginScreen from './views/LoginScreen';
+import HomeScreen from './views/HomeScreen/HomeScreen';
+import LoadingApp from './views/Loading/LoadingApp';
+import LoginScreen from './views/Login/LoginScreen';
 
 const Stack = createStackNavigator();
 const Drawer = createDrawerNavigator();
