+++ components/views/Proveedores/ProveedoresHeader.jsx
@@ -1,0 +1,33 @@
+import React from 'react';
+import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
+import { Ionicons } from '@expo/vector-icons';
+
+const ProveedoresHeader = ({ onMenu }) => (
+    <View style={styles.header}>
+        <TouchableOpacity onPress={onMenu}>
+            <Ionicons name="menu-outline" size={28} color="#333" />
+        </TouchableOpacity>
+        <Text style={styles.title}>Proveedores</Text>
+        <View style={{ width: 28 }} />
+    </View>
+);
+
+export default ProveedoresHeader;
+
+const styles = StyleSheet.create({
+    header: {
+        flexDirection: 'row',
+        alignItems: 'center',
+        justifyContent: 'space-between',
+        padding: 20,
+        backgroundColor: '#fff',
+        borderBottomWidth: 1,
+        borderBottomColor: '#e5e7eb',
+        elevation: 3,
+    },
+    title: {
+        fontSize: 20,
+        fontWeight: 'bold',
+        color: '#111827',
+    },
+});
