+++ components/views/Usuarios/FabButton.jsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import { TouchableOpacity, StyleSheet } from 'react-native';
+import { Ionicons } from '@expo/vector-icons';
+
+const FabButton = ({ onPress }) => (
+    <TouchableOpacity style={styles.fab} onPress={onPress}>
+        <Ionicons name="add" size={28} color="#fff" />
+    </TouchableOpacity>
+);
+
+const styles = StyleSheet.create({
+    fab: {
+        position: 'absolute',
+        bottom: 30,
+        right: 30,
+        backgroundColor: '#2563eb',
+        width: 60,
+        height: 60,
+        borderRadius: 30,
+        justifyContent: 'center',
+        alignItems: 'center',
+        elevation: 6,
+        shadowColor: '#000',
+        shadowOpacity: 0.2,
+        shadowOffset: { width: 0, height: 3 },
+        shadowRadius: 4,
+    },
+});
+
+export default FabButton;
